Try to factor the observaiton space
