+++ src/coreclr/scripts/cse_ml/jitml/constants.py
@@ -1,10 +1,7 @@
 """Constants and parameters for the project."""
 
-JITTYPE_ONEHOT_SIZE = 6
-
-BOOLEAN_FEATURES = JITTYPE_ONEHOT_SIZE + 7
-FLOAT_FEATURES = 9
-FEATURES = BOOLEAN_FEATURES + FLOAT_FEATURES
+MIN_CSE = 3
+MAX_CSE = 16
 
 REWARD_SCALE = 10
 REWARD_MIN = -1.0
